Working Behavior Trees and implemented roadrunner actions.

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/ftc7083/opmode/test/LimelightCalibrate.java
@@ -28,7 +28,7 @@ public class LimelightCalibrate  extends OpMode {
     @Override
     public void loop() {
 
-        double xDistance = (double) limelight.getDistance(Limelight.TargetPosition.SUBMERSIBLE);
+        Double xDistance = limelight.getDistance(Limelight.TargetPosition.SUBMERSIBLE);
 
         telemetry.addData("Distance to Target:",xDistance);
         telemetry.update();
